added product validation with joi

+++ api/Private/services/productService.js
@@ -1,5 +1,6 @@
 import db from '../../src/models';
 import lang from '../../language';
+import { validateCreateProduct, validateDeleteProduct } from '../validation/productValidation';
 
 class productService {
 
@@ -11,21 +12,25 @@ class productService {
 				return { message: lang(language).Product.getProducts.false, type: false };
 			}
 			return {
-				status: 200,
 				data: getProductResult,
 				message: lang(language).Product.getProducts.true,
 				type: true
 			};
 		}
 		catch (error) {
-			return {type: false, message: error.message};
+			return { type: false, message: error.message };
 		}
 	}
 	static async productAdd(req) {
-		//const language = req.decoded.language;
+		const language = req.decoded.language;
 		try {
 			const body = req.body;
 			const product = await db.products.create(body);
+
+			const validated_product = validateCreateProduct(product);
+			if (!validated_product) {
+				return { message: validated_product.message, type: false };
+			}
 			if (!product) {
 				const result = {
 					type: false,
@@ -34,21 +39,24 @@ class productService {
 				return result;
 			}
 			const result = {
-				type: true,
 				data: product,
-				message: 'product is created successfully'
+				message: lang(language).Product.productAdd.true,
+				type: true
 			};
 			return result;
 		}
 		catch (error) {
-			return {type: false, message: error.message};
+			return { type: false, message: error.message };
 		}
 	}
 	static async productFindById(req) {
 		const language = req.decoded.language;
 		try {
 			const productID = await db.products.findOne({ where: { id: req.params.id } });
-			return { data: productID, message: lang(language).Product.productFindById.true, type: true };
+			if (!productID)
+				return ({ type: false, message: lang(language).Product.productFindById.false });
+			else
+				return { data: productID, message: lang(language).Product.productFindById.true, type: true };
 		}
 		catch (error) {
 			return { message: lang(language).Product.productFindById.false, type: false };
@@ -57,18 +65,22 @@ class productService {
 	static async deleteProduct(req) {
 		const language = req.decoded.language;
 		try {
-			const product = await db.products.destroy({
-				where: {
-					id: req.params.id
-				}
-			});
-			if (!product) { 
-				const result = {
-					type: true,
-					message: 'Product is successfully deleted.'
-				};
-				return result;
+			const product = await db.products.findOne({ where: { id: req.params.id } });
+			const validated_product = validateDeleteProduct(product);
+			if (!validated_product) {
+				return { message: validated_product.message, type: false };
 			}
+			if (product) {
+				const productID = await db.products.destroy({
+					where: {
+						id: req.params.id
+					}
+				});
+				if (productID)
+					return ({ type: true, message: 'product is deleted' });
+			}
+			else 
+				return ({  message: lang(language).Product.deleteProduct.false, type: false });
 		}
 		catch (error) {
 			return { message: lang(language).Product.deleteProduct.false, type: false };
